% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePhenoTable.R
\name{makePhenoTable}
\alias{makePhenoTable}
\title{Converts narrow code table to wide phenotype table}
\usage{
makePhenoTable(data, code_var = "dx", vocab = "ICD")
}
\arguments{
\item{data}{Data frame/table with which to apply function to}

\item{code_var}{String. Specifies the column name of the code variable}

\item{vocab}{String. Specifies which clinical vocabulary is in use (ICD10, OPCS4, CPRD)}
}
\value{
data.table which is a wide phenotype table
}
\description{
Converts narrow code table to wide phenotype table
}
