% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesisEHR.R
\name{synthesisEHR}
\alias{synthesisEHR}
\title{Generates synthetic EHR data from imported phenotypes for testing}
\usage{
synthesisEHR(n, vocab = "ICD")
}
\arguments{
\item{n}{number of patients}

\item{vocab}{A string - specifies which clinical vocabulary is in use (ICD10, OPCS4, CPRD)}
}
\value{
data.table
}
\description{
Generates synthetic EHR data from imported phenotypes for testing
}
